+++ kob-backend/src/main/java/com/sdu/kob/consumer/WebSocketServer.java
@@ -31,10 +31,10 @@ public class WebSocketServer {
     public static final String removePlayerUrl = "http://127.0.0.1:3001/go/matching/remove/";
     private static final String setEngineUrl = "http://8.142.10.225:5001/set";
 
-    final public static ConcurrentHashMap<Integer, WebSocketServer> goUsers = new ConcurrentHashMap<>();
-    final public static ConcurrentHashMap<Integer, String> user2room = new ConcurrentHashMap<>();
+    final public static ConcurrentHashMap<Long, WebSocketServer> goUsers = new ConcurrentHashMap<>();
+    final public static ConcurrentHashMap<Long, String> user2room = new ConcurrentHashMap<>();
     final public static ConcurrentHashMap<String, Room> rooms = new ConcurrentHashMap<>();
-    final public static CopyOnWriteArraySet<Integer> matchingUsers = new CopyOnWriteArraySet<>();
+    final public static CopyOnWriteArraySet<Long> matchingUsers = new CopyOnWriteArraySet<>();
 
     private User user;
     private Session session = null;
@@ -63,7 +63,7 @@ public class WebSocketServer {
         // 建立连接 所有与连接相关的信息都会存到这个类中
         this.session = session;
         // 1. 从token中读取建立连接的用户是谁 拿到id
-        int userId = JwtUtil.JWTAuthentication(token);
+        long userId = JwtUtil.JWTAuthentication(token);
         // 2. 根据id查找用户
         this.user = userDAO.findById(userId);
         // 3. 将用户存下来
@@ -78,7 +78,7 @@ public class WebSocketServer {
     @OnClose
     public void onClose() {
         if (this.user != null) {
-            Integer userId = this.user.getId();
+            Long userId = this.user.getId();
             goUsers.remove(userId);
             System.out.println("Closed!");
         }
@@ -95,7 +95,7 @@ public class WebSocketServer {
             cancelMatching();
         } else if ("play".equals(event)) {
             Integer x = data.getInteger("x"), y = data.getInteger("y");
-            Integer userId = this.user.getId();
+            Long userId = this.user.getId();
             if (x == -1 && y == -1) {
                 if (Objects.equals(userId, rooms.get(user2room.get(userId)).blackPlayer.getId())) {
                     rooms.get(user2room.get(userId)).setLoser(BLACK);
@@ -104,8 +104,8 @@ public class WebSocketServer {
             }
             rooms.get(user2room.get(userId)).setNextStep(x, y, false);
         } else if ("request_play".equals(event)) {
-            Integer friendId = data.getInteger("friend_id");
-            Integer requestId = data.getInteger("request_id");
+            Long friendId = data.getLong("friend_id");
+            Long requestId = data.getLong("request_id");
             sendRequest2play(friendId, requestId);
         } else if ("request_draw".equals(event)) {
             String roomId = user2room.get(user.getId());
@@ -115,25 +115,25 @@ public class WebSocketServer {
                 playRoom.setNextStep(-2, -2, false);
                 return;
             }
-            Integer friendId = data.getInteger("friend_id");
+            Long friendId = data.getLong("friend_id");
             sendRequest2DrawOrRegret(friendId, 1);
         } else if ("request_cancel".equals(event)) {
-            Integer friendId = data.getInteger("friend_id");
+            Long friendId = data.getLong("friend_id");
             sendRequest2Cancel(friendId);
         } else if ("refuse_invitation".equals(event)) {
-            Integer friendId = data.getInteger("friend_id");
+            Long friendId = data.getLong("friend_id");
             sendRefuseMessage(friendId);
         } else if ("accept_invitation".equals(event)) {
-            Integer aId = data.getInteger("user_id");
-            Integer bId = data.getInteger("friend_id");
+            Long aId = data.getLong("user_id");
+            Long bId = data.getLong("friend_id");
             getReady(aId, bId);
             startGame(aId, bId);
         } else if ("accept_draw".equals(event)){
-            Integer aId = data.getInteger("user_id");
+            Long aId = data.getLong("user_id");
             rooms.get(user2room.get(aId)).setLoser(-2);
             rooms.get(user2room.get(aId)).setNextStep(-2, -2, false);
         } else if ("engine_play".equals(event)) {
-            Integer userId = data.getInteger("user_id");
+            Long userId = data.getLong("user_id");
             Integer level = data.getInteger("level");
             startAIPlaying(userId, level);
         } else if ("request_regret".equals(event)) {
@@ -143,10 +143,10 @@ public class WebSocketServer {
                 playRoom.regretPlay(data.getInteger("which"));
                 return;
             }
-            Integer friendId = data.getInteger("friend_id");
+            Long friendId = data.getLong("friend_id");
             sendRequest2DrawOrRegret(friendId, 2);
         } else if ("accept_regret".equals(event)) {
-            Integer aId = data.getInteger("user_id");
+            Long aId = data.getLong("user_id");
             Integer which = data.getInteger("which");
             rooms.get(user2room.get(aId)).regretPlay(which);
         }
@@ -157,7 +157,7 @@ public class WebSocketServer {
         error.printStackTrace();
     }
 
-    private void getReady(Integer aId, Integer bId) {
+    private void getReady(Long aId, Long bId) {
         WebSocketServer aClient = goUsers.get(aId);
         WebSocketServer bClient = goUsers.get(bId);
         JSONObject resp = new JSONObject();
@@ -190,7 +190,7 @@ public class WebSocketServer {
      * @param message 消息
      */
     public static void sendGroupMessage(String message) {
-        for (Map.Entry<Integer, WebSocketServer> entry : goUsers.entrySet()) {
+        for (Map.Entry<Long, WebSocketServer> entry : goUsers.entrySet()) {
             entry.getValue().sendMessage(message);
         }
     }
@@ -199,7 +199,7 @@ public class WebSocketServer {
      * 拒绝请求
      * @param friendId 被拒绝的用户id
      */
-    private void sendRefuseMessage(Integer friendId) {
+    private void sendRefuseMessage(Long friendId) {
         if (goUsers.get(friendId) == null) {
             matchingUsers.remove(this.user.getId());
             return;
@@ -213,7 +213,7 @@ public class WebSocketServer {
     }
 
     // 取消发送请求
-    private void sendRequest2Cancel(Integer friendId) {
+    private void sendRequest2Cancel(Long friendId) {
         if (goUsers.get(friendId) == null) {
             matchingUsers.remove(this.user.getId());
             return;
@@ -231,7 +231,7 @@ public class WebSocketServer {
      * @param opponentId 对手的id
      * @param type 类型：1和棋 2悔棋
      */
-    private void sendRequest2DrawOrRegret(Integer opponentId, int type) {
+    private void sendRequest2DrawOrRegret(Long opponentId, int type) {
         if (goUsers.get(opponentId) == null) {
             return;
         }
@@ -246,14 +246,14 @@ public class WebSocketServer {
      * 向另一名玩家发送对战请求
      * @param friendId 被邀请人的id
      */
-    private void sendRequest2play(Integer friendId, Integer requestId) {
+    private void sendRequest2play(Long friendId, Long requestId) {
         if (goUsers.get(friendId) != null) {
             matchingUsers.add(requestId);
             matchingUsers.add(friendId);
             WebSocketServer friendClient = goUsers.get(friendId);
             JSONObject resp = new JSONObject();
             JSONObject request_user = new JSONObject();
-            User requestUser = userDAO.findById((int)requestId);
+            User requestUser = userDAO.findById((long)requestId);
             resp.put("event", "request_play");
             request_user.put("id", requestUser.getId());
             request_user.put("username", requestUser.getUserName());
@@ -290,11 +290,11 @@ public class WebSocketServer {
         restTemplate.postForObject(removePlayerUrl, cancelMatchData, String.class);
     }
 
-    private void startAIPlaying(Integer userId, Integer level) {
+    private void startAIPlaying(Long userId, Integer level) {
         JSONObject respGame = new JSONObject();
-        User human = userDAO.findById((int) userId);
+        User human = userDAO.findById((long) userId);
         User engine = new User(level);
-        Room room = new Room(19, 19, userId, human, -1, engine, true);
+        Room room = new Room(19, 19, userId, human, -1L, engine, true);
         user2room.put(userId, room.uuid);
         rooms.put(room.uuid, room);
 
@@ -332,17 +332,17 @@ public class WebSocketServer {
     }
 
     // 开始下棋
-    public static void startGame(Integer aId, Integer bId) {
+    public static void startGame(Long aId, Long bId) {
         matchingUsers.remove(aId);
         matchingUsers.remove(bId);
-        User a = userDAO.findById((int) aId);
-        User b = userDAO.findById((int) bId);
+        User a = userDAO.findById((long) aId);
+        User b = userDAO.findById((long) bId);
 
         JSONObject respGame = new JSONObject();
         Random random = new Random();
         int seed = random.nextInt();
         // temp
-        Integer blackId, whiteId;
+        Long blackId, whiteId;
         Room room;
         if (seed * 10 >= 5) {
             blackId = aId;
