+++ kob-backend/src/main/java/com/sdu/kob/controller/UserController.java
@@ -26,7 +26,7 @@ public class UserController {
 
     @RequestMapping(value = "/api/user/getFollowedAndFollowersCount/", method = RequestMethod.POST)
     public JSONObject getFollowedAndFollowersCountAndGuests(@RequestParam Map<String, String> data, Principal principal) {
-        Integer userId = Integer.parseInt(data.get("user_id"));
+        Long userId = Long.parseLong(data.get("user_id"));
         String requestUserName = principal.getName();
         return userService.getFollowedAndFollowersCountAndGuests(userId, requestUserName);
     }
