+++ kob-backend/src/main/java/com/sdu/kob/repository/PostDAO.java
@@ -10,13 +10,13 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.Date;
 import java.util.List;
 
-public interface PostDAO extends JpaRepository<Post, Integer> {
-    List<Post> findByUserId(Integer userId);
+public interface PostDAO extends JpaRepository<Post, Long> {
+    List<Post> findByUserId(Long userId);
 
-    Post findById(int id);
+    Post findById(long id);
 
     @Modifying
     @Transactional
     @Query(value = "update post set user_id = ?2, title = ?3, content = ?4, modify_time = ?5 where id = ?1", nativeQuery = true)
-    void updatePost(int id, int UserId, String title, String content, Date modifyTime);
+    void updatePost(long id, long UserId, String title, String content, Date modifyTime);
 }
