+++ kob-backend/src/main/java/com/sdu/kob/service/RoomService.java
@@ -5,7 +5,7 @@ import com.alibaba.fastjson.JSONObject;
 public interface RoomService {
     JSONObject getUsersInRoom(String roomId);
 
-    String leaveRoom(String roomId, Integer userId);
+    String leaveRoom(String roomId, Long userId);
 
-    JSONObject getBoardInRoom(Integer userId, String roomId);
+    JSONObject getBoardInRoom(Long userId, String roomId);
 }
